Move hipblaslt_op from compile_hipblaslt.cpp to compile_hipblaslt.hpp
