+++ src/targets/gpu/compile_hipblaslt.cpp
@@ -29,37 +29,10 @@
 #include <migraphx/iterator_for.hpp>
 #include <migraphx/instruction.hpp>
 #include <migraphx/make_op.hpp>
-#include <migraphx/register_op.hpp>
-#include <migraphx/op/identity.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 namespace gpu {
-
-struct hipblaslt_op
-{
-    operation op = op::identity{};
-
-    template <class Self, class F>
-    static auto reflect(Self& self, F f)
-    {
-        return pack(f(self.op, "op"));
-    }
-
-    std::string name() const { return "gpu::hipblaslt_op"; }
-
-    shape compute_shape(std::vector<shape> inputs) const
-    {
-        inputs.push_back(inputs.back());
-        return op.compute_shape(inputs);
-    }
-
-    std::ptrdiff_t output_alias(const std::vector<shape>& shapes) const
-    {
-        return shapes.size() - 1;
-    }
-};
-MIGRAPHX_REGISTER_OP(hipblaslt_op);
 
 static size_t compile(migraphx::context& ctx, operation& op, instruction_ref ins)
 {
